% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{analysePlateFeatures}
\alias{analysePlateFeatures}
\title{Analyses the per-plate features}
\usage{
analysePlateFeatures(platenames, configdir, grouping, correlation_threshold)
}
\arguments{
\item{platenames}{A list of platenames to use as biological replicates for the analysis. This list can be one element long}

\item{configdir}{The directory of the configuration files}

\item{grouping}{Can take the values "drugs", "combo", or "type". "Drugs" indicates that the drugs alone should be used to label the wells, 
"combo" indicates that the combination of drugs and concentration should be used, and "type" indicates that the treatment type should be used
("drug", "positive control", or "negative control")}

\item{correlation_threshold}{Only features, which correlate between replicates are used for the analysis. This threshold indicates how strongly
they should correlate}
}
\value{
A boolean value indicating if the function was successful
}
\description{
Performs quality control on the per-plate features followed by a a shrinkage linear discriminant analysis (bioconductor package 'sda').
If there are multiple plates, use these as replicates. If there is only one plate, however, use wells with identical treatment on the plate as
biological replicates. The current implementation requires that there is an equal number of wells for each drug treatment
}
\examples{
print(analysePlateFeatures)
}
\author{
Jan Sauer
}
