% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{calcFeaturesPerPlate}
\alias{calcFeaturesPerPlate}
\title{Calculates the mean feature vector for an entire plate}
\usage{
calcFeaturesPerPlate(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{platename}{The plate that the well is on}

\item{row}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{col}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the function was successful
}
\description{
Runs calcFeaturesPerWell() for each well and saves it as a file for the plate
}
\examples{
print(calcFeaturesPerPlate)
}
\author{
Jan Sauer
}
