% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{calcFeaturesPerWell}
\alias{calcFeaturesPerWell}
\title{Calculates the mean feature vector for an entire well}
\usage{
calcFeaturesPerWell(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
The combined feature vector of the well
}
\description{
Calculates the mean feature vector as well as the standard deviation of all features over all organoids for a single well.
Also determines the mean number of cells in each organoid as well as how many organoids and single cells there are.
}
\examples{
print(calcFeaturesPerWell)
}
\author{
Jan Sauer
}
