% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityDistribution.R
\name{calcIntensityDistribution}
\alias{calcIntensityDistribution}
\title{Calculate Intensity Distribution}
\usage{
calcIntensityDistribution(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the microscope (Unnecessary, only here for consistent signature)}

\item{platename}{The ID (and folder name) of the plate}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the save was successful
}
\description{
Calculate intensity distribution for a well
}
\examples{
print(calcIntensityDistribution)
}
\author{
Jan Sauer
}
