% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLogs.R
\name{checkLogs}
\alias{checkLogs}
\title{Check logfiles of jobs}
\usage{
checkLogs(plateIndir, platename, outfile, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the 
microscope}

\item{platename}{The ID (and folder name) of the plate}

\item{outfile}{The location to write the fix-script to}

\item{configdir}{The configuration file, which defines the number of 
z-stacks, channels, and fields}
}
\value{
A list of lists indicating incorrect hdf5dir files and projections
}
\description{
This function checks whether jobs ran successfully. It checks 
  log files and verifies the hdf5dir and hdf5projection entry for all files 
  with non-empty log files. Note that it only checks the hdf5validation 
  file to confirm the file in hdf5dir, it does not actually compare TIF 
  imagtes with the h5 files. The function then creates shell scripts to fix 
  the bad files.
}
\examples{
print(checkLogs)
}
\author{
Jan Sauer
}
