% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{cleanupSegmap}
\alias{cleanupSegmap}
\title{Clean up segmentation mask}
\usage{
cleanupSegmap(segmap)
}
\arguments{
\item{segmap}{A single 2D matrix with integer values}
}
\value{
A named list of 2D matrices. Currently with the entries "organoids" 
and "other".
}
\description{
"Cleans up" a segmentation mask. This means that two separate 
masks will be created: one mask containing all closed organoids and one 
mask containing all foreground and edge fragments not inside these closed 
organoids. Features for these two classes of items are calculated 
separately.
}
\examples{
print(cleanupSegmap)
}
\author{
Jan Sauer
}
