% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{combineFeatures}
\alias{combineFeatures}
\title{Combines the features per well into a single file for an entire plate}
\usage{
combineFeatures(platename, configdir)
}
\arguments{
\item{platename}{The plate that the well is on}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the combining was successful
}
\description{
Combines all features for each individual well of a plate into a single file. This function requires that 
features for well A01 exist for the extraction of feature names.
}
\examples{
print(combineFeatures)
}
\author{
Jan Sauer
}
