% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{extractOrganoidFeatures}
\alias{extractOrganoidFeatures}
\title{Extract organoid features}
\usage{
extractOrganoidFeatures(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the features were correctly calculated and saved
}
\description{
Loads a projection and the corresponding segmentation map and 
calculates the full set of EBImage features as well as the texture features 
of only the primitive foreground / background segmentation
}
\examples{
print(extractOrganoidFeatures)
}
\author{
Jan Sauer
}
