% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{extractOrganoidFeatures_texture}
\alias{extractOrganoidFeatures_texture}
\title{Extract organoid features}
\usage{
extractOrganoidFeatures_texture(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the features were correctly calculated and saved
}
\description{
Loads a projection and the corresponding segmentation map and 
calculates the haralick features using EBImage. The organoids are not 
differentiated individually
}
\examples{
print(extractOrganoidFeatures_texture)
}
\author{
Jan Sauer
}
