% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{extractOrganoidFeatures_v1}
\alias{extractOrganoidFeatures_v1}
\title{Extract organoid features}
\usage{
extractOrganoidFeatures_v1(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{This parameter does nothing, but makes it easier to integrate the function into the workflow}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the features were correctly calculated and saved
}
\description{
!THIS METHOD IS OBSOLETE! Segments the projected images via thresholding to obtain a rough estimate of the organoids and then calculates the 
features (EBImage::computeFeatures) for each individual organoid. Both the mask and the features are saved as an hdf5 file for each well.
}
\examples{
print(extractOrganoidFeatures_v1)
}
\author{
Jan Sauer
}
