% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{featuresHdf5Filename}
\alias{featuresHdf5Filename}
\title{Create hdf5 filename for features}
\usage{
featuresHdf5Filename(filedir, platename, row, col, configdir = NULL)
}
\arguments{
\item{filedir}{The directory of the file}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
The hdf5 filename
}
\description{
Creates hdf5 filenames with given parameters for the organoid features
}
\examples{
print(featuresHdf5Filename)
}
\author{
Jan Sauer
}
