% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{featuresMaskHdf5Filename}
\alias{featuresMaskHdf5Filename}
\title{Create hdf5 filename for binary masks}
\usage{
featuresMaskHdf5Filename(filedir, platename, row, col, configdir)
}
\arguments{
\item{filedir}{The directory of the file}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}
}
\value{
The hdf5 filename
}
\description{
Creates hdf5 filenames with given parameters for the binary masks used to segment the projected images
}
\examples{
print(featuresMaskHdf5Filename)
}
\author{
Jan Sauer
}
