% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{file.isopen}
\alias{file.isopen}
\title{Test if files are open}
\usage{
file.isopen(fn)
}
\arguments{
\item{fn}{A vector of filenames}
}
\value{
Logical indicating for each file, if it is open
}
\description{
Tests for each file, if it is open
}
\examples{
print(file.isopen)
}
\author{
Bernd Fischer
}
