% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{filesComplete}
\alias{filesComplete}
\title{Test if all files of one well are complete}
\usage{
filesComplete(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{The base directory on the server, which contains a folder for each plate}

\item{platename}{The ID (and folder name) of the plate}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The configuration file, which defines all the directories, the number of z-stacks, channels, and fields}
}
\value{
Logical indicating if all files are complete
}
\description{
Tests if all required files for processing one well are 
existing and not open any more.
}
\examples{
print(filesComplete)
}
\author{
Bernd Fischer
}
