% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeatures.R
\name{get_all_compounds}
\alias{get_all_compounds}
\title{Get a list of all compounds}
\usage{
get_all_compounds(configdir)
}
\arguments{
\item{configdir}{The directory of the configuration files}
}
\value{
A vector of strings
}
\description{
Get a list of all compounds
}
\author{
Jan Sauer

#@examples print(get_all_compounds)
#@export
}
