% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeatures.R
\name{get_features_by_compound}
\alias{get_features_by_compound}
\title{Get feature vector by compound}
\usage{
get_features_by_compound(compound, organoid_type, feature_type, configdir)
}
\arguments{
\item{compound}{The compound for which to retrieve the feature vectors}

\item{organoid_type}{Either "mouse", "human", or "both", depending on which 
organoid data to search through}

\item{feature_type}{The subfolder of the "features" PROMISE folder from 
which to retrieve the features (e.g. "inception" for 
inception features)}

\item{configdir}{The directory of the configuration files}
}
\value{
A matrix of features. Each row represents one well, each column 
        represents a feature. The rownames indicate the source well
}
\description{
Retrieve all feature vectors corresponding to a compound. This 
             searches through all plates currently found in the "featuresdir"
             location
}
\author{
Jan Sauer

# @examples print(get_features_by_compound)
# @export
}
