% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_api.R
\name{get_features_by_well}
\alias{get_features_by_well}
\title{Get feature vector by a specific well}
\usage{
get_features_by_well(platename, row, col, feature_type, configdir)
}
\arguments{
\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{feature_type}{The subfolder of the "features" PROMISE folder from 
which to retrieve the features (e.g. "inception" for 
inception features)}

\item{configdir}{The directory of the configuration files}
}
\value{
A feature vector
}
\description{
Retrieves the features for a given well
}
\examples{
print(get_features_by_well)
}
\author{
Jan Sauer
}
