% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{imageFilenames}
\alias{imageFilenames}
\title{Get filenames associated with a specific well}
\usage{
imageFilenames(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the microscope}

\item{platename}{The ID (and folder name) of the plate}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory with configuration files}
}
\value{
A string containing the list of existing files for the given well
}
\description{
This function finds all (existing) files associated with a specific well based on the number of z-stacks, channels, and fields defined in a configuration file
}
\examples{
print(imageFilenames)
}
\author{
Jan Sauer
}
