% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeatures.R
\name{loadFeatures}
\alias{loadFeatures}
\title{Load all features for a plate}
\usage{
loadFeatures(platename, configdir, feature_type)
}
\arguments{
\item{platename}{The plate ID}

\item{configdir}{The directory of the configuration files}

\item{feature_type}{The type of feature to return. See details.}
}
\value{
A data.frame of features. Each row represents one well's field, 
        each column represents a feature.
}
\description{
Retrieve all feature for a plate
}
\details{
The feature_type parameter can have the following values:
\describe{
  \item{organoids}{Returns features for individual organoids. Touching 
  organoids are separated as best as possible.}
  \item{clumps}{Returns features for individual clumps. A "clump" is any 
  contiguous foreground patch without any object separation.} 
  \item{foreground}{Returns a single feature set for the entire foreground 
  of an image without any object separation. This feature set only includes 
  shape-independent haralick features}}
}
\examples{
print(loadFeatures)
}
\author{
Jan Sauer
}
