% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFeatures.R
\name{loadFeaturesForWell}
\alias{loadFeaturesForWell}
\title{Load features for a well}
\usage{
loadFeaturesForWell(platename, row, col, configdir)
}
\arguments{
\item{platename}{The plate ID}

\item{row}{The plate row}

\item{col}{The plate column}

\item{configdir}{The directory of the configuration files}
}
\value{
A list of feature dataframes. The entries are as follows: 
  "organoid_features" correspond to features for individual organoids.
  "clump_features" corespond to features of entire contiguous foreground 
  patches without any (i.e. groups of connected organoids)
  "texture_features" correspond to the shape-independent haralick features 
  of the foreground without any object separation.
}
\description{
Retrieve all features for a well
}
\examples{
print(loadFeaturesForWell)
}
\author{
Jan Sauer
}
