% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLibrary.R
\name{loadLibrary}
\alias{loadLibrary}
\title{Load layout library}
\usage{
loadLibrary(library, configdir, cols = c("Well_ID_384", "Product.Name",
  "concentration"))
}
\arguments{
\item{library}{The library ID (e.g. L02)}

\item{configdir}{The directory of the configuration files}

\item{cols}{A custom vector of columns to return. Column names not found 
in the library file are silently ignored.}
}
\value{
A library dataframe
}
\description{
Loads the layout library
}
\examples{
print(loadLibrary)
}
\author{
Jan Sauer
}
