% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLibrary.R
\name{loadLibraryFromPlate}
\alias{loadLibraryFromPlate}
\title{Load layout library fron Plate}
\usage{
loadLibraryFromPlate(platename, configdir, cols = c("Well_ID_384",
  "Product.Name", "concentration"))
}
\arguments{
\item{platename}{The plate ID}

\item{configdir}{The directory of the configuration files}

\item{cols}{A custom vector of columns to return. Column names not found 
in the library file are silently ignored.}
}
\value{
A library dataframe
}
\description{
Loads the layout library corresponding to the plate
}
\examples{
print(loadLibraryFromPlate)
}
\author{
Jan Sauer
}
