% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5projectionApi.R
\name{loadProjection}
\alias{loadProjection}
\title{Load or save a projection from hdf5 to an image}
\usage{
loadProjection(platename, row, col, configdir, outfile = NULL,
  preprocess = TRUE)
}
\arguments{
\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}

\item{outfile}{The file to write to. Leave as NULL to not save}

\item{preprocess}{A boolean indicating if the intensities should be 
preprocessed.}
}
\value{
A 4D array with the dimensions (X, Y, channel, field)
}
\description{
Loads the hdf5projection file for a given well and returns it 
as an array and optionally saves it to a TIF file. The fields are saved as 
a single image. Note that there is NO GUARANTEE that the TIFF images have 
the true intensities. Analysis should be performed exclusively on the hdf5 
files. 

The function can optionally preprocess the intensities. The raw images are 
most likely too dark to see the the naked eye. To fix this, the brightest
0.1% of all pixels can be cropped for each channel independently. This 
makes the picture more easily viewable with the naked eye. This 
preprocessing only applies to the saved file. The function returns the raw 
integer values in the hdf5 file.
}
\examples{
print(loadProjection)
}
\author{
Jan Sauer
}
