% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5projectionApi.R
\name{loadSegmentationMask}
\alias{loadSegmentationMask}
\title{Load or save a segmentation mask from hdf5 to an image}
\usage{
loadSegmentationMask(platename, row, col, configdir, outfile = NULL)
}
\arguments{
\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files}

\item{outfile}{The file to write to. Leave as NULL to not save}
}
\value{
A 3D array with the dimensions (X, Y, field)
}
\description{
Loads the segmentation mask for a given well and returns it 
as an array and optionally saves it to a TIF file. The fields are saved as 
a single image.
}
\examples{
print(loadSegmentationMask)
}
\author{
Jan Sauer
}
