% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnails.R
\name{makeIndexPage}
\alias{makeIndexPage}
\title{Writes an index html-page for the thumbnail images of z-projecion}
\usage{
makeIndexPage(platename, configdir)
}
\arguments{
\item{platename}{The ID (and folder name) of the plate}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the index file was written successfully
}
\description{
This function writes an index.html page for the thumbnail images of the z-projection
}
\examples{
print(makeIndexPage)
}
\author{
Bernd Fischer
}
