% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnails.R
\name{makeIndexPage_segmentation}
\alias{makeIndexPage_segmentation}
\title{Writes an index html-page for the thumbnail images of the DNN segmentation}
\usage{
makeIndexPage_segmentation(platename, configdir)
}
\arguments{
\item{platename}{The ID (and folder name) of the plate}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the index file was written successfully
}
\description{
This function writes an index.html page for the thumbnail images of the z-projection
}
\examples{
print(makeIndexPage_segmentation)
}
\author{
Jan Sauer
}
