% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tifdiagnostics.R
\name{moveTifDiagnostics}
\alias{moveTifDiagnostics}
\title{TIF Diagnostics}
\usage{
moveTifDiagnostics(plateIndir, platename, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the microscope}

\item{platename}{The ID (and folder name) of the plate}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\description{
This extracts all files that are not well images and places them in a diagnostics directory
}
\examples{
print(moveTifDiagnostics)
}
\author{
Jan Sauer
}
