
\name{phPatt}
\alias{phPatt}
\title{Phenotype Pattern Definition Set}
\description{  
    This hypothetical phenotype pattern definition set \emph{phPatt} has three columns: stat.coef, stat.func, and endpt.vars. 
    It defines an associatin pattern for three phenotypes.
}
\usage{
data(phPatt)
}
\keyword{misc}


