% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{processWell}
\alias{processWell}
\title{Submits a job to the cluster queue to process a single well}
\usage{
processWell()
}
\value{
Nothing is returned
}
\description{
This function is called by the batch script processWell.sh.
 It submits a single well process to the cluster queue.
}
\examples{
print(processWell)
}
\author{
Bernd Fischer
}
