% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zprojection.R
\name{projection}
\alias{projection}
\title{Project all images in a well}
\usage{
projection(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{The base directory on the server, which contains a folder for each plate}

\item{platename}{The ID (and folder name) of the plate}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the projection was successful
}
\description{
This function projects all images in a well and stores the results in an hdf5 file
}
\examples{
print(projection)
}
\author{
Jan Sauer
}
