% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\docType{package}
\name{promise-package}
\alias{promise-package}
\title{promise-package}
\description{
A data package for the PROMISE database.
}
\details{
Use \code{data(package='promise')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{promise}
}
