% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\docType{package}
\name{promise}
\alias{promise}
\alias{promise-package}
\alias{promise-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
promise
A data package for promise.

Use \code{data(package='promise')$results[, 3]} tosee a list of availabledata sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{promise}
}
