% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTIFF.R
\name{renameTIFF}
\alias{renameTIFF}
\title{Rename TIFF files to the expected format}
\usage{
renameTIFF(plateIndir, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the microscope}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the rename was successful
}
\description{
Rename all TIFF files of in a plate folder to the expected format
}
\examples{
print(renameTIFF)
}
\author{
Jan Sauer
}
