
\name{sampExprSet}
\alias{sampExprSet}
\title{An Example Expression Set}
\description{  
    This hypothetical expression set \emph{sampExpSet} belongs to an \emph{ExpressionSet} class. 
    It contains 100 genomic features (probe_1 to probe_100) for 50 subjects (array_1 to array_50) 
    and phenotype data of drugLevel, residualDisease, obsTime, obsCensor and strat. The expression values can be accessed by \emph{exprs(sampExprSet)}.
    The phenotype data can be accessed by \emph{pData(phenoData(sampExprSet))}
}
\usage{
data(sampExprSet)
}
\keyword{misc}
