% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{set_data_path}
\alias{set_data_path}
\title{Set promise data path}
\usage{
set_data_path(path = NULL, .interactive = TRUE, .update = FALSE)
}
\arguments{
\item{path}{A \code{character} path to and including name and extension of the .rda file.}

\item{.interactive}{\code{logical} Use file browser}

\item{.update}{A \code{logical}, for whether or not to update symlink.}
}
\description{
Setting up symlink for promise data
}
\details{
Since promise as of date contains PHI data processed data can't be bundled with
the package. If you mount the box.com folder as a local drive, one can symlink
the data to the promise data folder.
}
\examples{
\dontrun{
#First non-interactive use:
set_data_path("C:/Users/usename/Box/jolo_projects/promise/data/promise_v.0.1.3.rda", .interactive = FALSE)

#For interactive use
set_data_path()

#For updating path
set_data_path(.update = TRUE)
}
}
