% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{setupWatchdog}
\alias{setupWatchdog}
\title{Setup the watchdog system}
\usage{
setupWatchdog(bindir = "~/bin", pythondir = file.path(getwd(),
  "config/segmentation_config"), hdf5dir = file.path(getwd(), "hdf5dir"),
  hdf5validationdir = file.path(getwd(), "hdf5validation"),
  hdf5projection = file.path(getwd(), "hdf5projection"),
  htmldir = file.path(getwd(), "htmldir"), layoutdir = file.path(getwd(),
  "layouts"), featuresdir = file.path(getwd(), "features"),
  segmentationdir = file.path(getwd(), "segmentation"),
  segmentationhtmldir = file.path(getwd(), "segmentationhtmldir"),
  tifdiagnosticsdir = file.path(getwd(), "tifdiagnostics"),
  configdir = file.path(getwd(), "configdir"), showWarnings = TRUE)
}
\arguments{
\item{bindir}{The bash scripts are written in this directory}

\item{pythondir}{The directory that contains the python script(s) for the segmentation as well as the weights and layout information for dnnSwift}

\item{hdf5dir}{The directory on the server for writing the HDF5 files containing the full images}

\item{hdf5validationdir}{The directory on the server for writing the status files indicating if the TIF images and the HDF5 files are identical}

\item{hdf5projection}{The directory on the server for writing the HDF5 files containing the z-projection}

\item{htmldir}{The directory on the server for writing thumbnail images}

\item{layoutdir}{The directory on the server for the plate layouts}

\item{featuresdir}{The directory on the server for the feature files}

\item{segmentationdir}{The directory in which the DNN segmentation is stored}

\item{segmentationhtmldir}{The directory in which the preview images for the segmentation are stored}

\item{tifdiagnosticsdir}{The directory in which TIF diagnostic files are stored}

\item{configdir}{The configuration file, which defines all the directories, the number of z-stacks, channels, and fields}

\item{showWarnings}{A boolean value indicating if a warning should be shown if the config file already exists when this function is executed}
}
\value{
Nothing is returned
}
\description{
This function generates a number of bach scripts and a template config file for the watchdog
}
\examples{
print(setupWatchdog)
}
\author{
Bernd Fischer
}
