% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{singleImageFilename}
\alias{singleImageFilename}
\title{Create tif image filename name}
\usage{
singleImageFilename(configdir, row, col, field, channel, stack)
}
\arguments{
\item{configdir}{The directory containing watchdogConfig.R}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{field}{The field of the well}

\item{channel}{The color channel}

\item{stack}{The z-stack of the image}
}
\value{
The image filename
}
\description{
Creates the tif image filename given with the parameters
}
\examples{
print(singleImageFilename)
}
\author{
Jan Sauer
}
