% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{startWorkflow}
\alias{startWorkflow}
\title{Read command line and start R function}
\usage{
startWorkflow()
}
\value{
Nothing is returned
}
\description{
This function is called by the watchdog to start the processing
 of a single well. It reads the arguments from command line and passes them
 to the function specified as its first argument. This function is only 
 usefull, if you want to process a single well by a shell command. You can
 do so by calling
 >R -e "library(PROMISE);startWorkflow --args tif2hdf5 plateIndir platename row col configdir".
}
\examples{
print(startWorkflow)
}
\author{
Bernd Fischer
}
