% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSubmission.R
\name{submitPlateToQueue}
\alias{submitPlateToQueue}
\title{A helper script to submit an entire plate to the queue}
\usage{
submitPlateToQueue(workflow, plateIndir, platename, configdir)
}
\arguments{
\item{workflow}{The workflow function to call (e.g. PROMISEconversion, PROMISEfeatures)}

\item{plateIndir}{The incoming directory for the plate. Can be a dummy value for workflows that don't need it.}

\item{platename}{The plate that the well is on}

\item{configdir}{The directory of the configuration files}
}
\value{
A boolean value indicating if the features were correctly calculated and saved
}
\description{
A helper script to submit an entire plate to the queue
}
\examples{
print(submitPlateToQueue)
}
\author{
Jan Sauer
}
