% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{thumbnailFilename}
\alias{thumbnailFilename}
\title{Create image filenames from row and col for thumbnail images}
\usage{
thumbnailFilename(filedir, platename, row, col, configdir = NULL, level = 1,
  addPath = TRUE)
}
\arguments{
\item{filedir}{The directory of the file}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The directory of the configuration files. Unnecessary but here for backwards compatibility}

\item{level}{1: The larger thumbnail; 2: The smaller thumbnail}

\item{addPath}{Logical, if TRUE, the full path is added to the filename}
}
\value{
The image filename
}
\description{
Creates image filenames with given parameters for thumbnail images
}
\examples{
print(thumbnailFilename)
}
\author{
Bernd Fischer
}
