% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateHDF5.R
\name{validateHDF5}
\alias{validateHDF5}
\title{Validates HDF5 files of raw image data}
\usage{
validateHDF5(plateIndir, platename, row, col, configdir)
}
\arguments{
\item{plateIndir}{The directory name of the plate as generated by the microscope}

\item{platename}{The ID (and folder name) of the plate}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{The configuration file, which defines the number of z-stacks, channels, and fields}
}
\value{
A boolean value indicating if the two files match
}
\description{
This function checks that the HDF5 file created for a well matches the raw image files
}
\examples{
print(validateHDF5)
}
\author{
Jan Sauer
}
