% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{validationHdf5Filename}
\alias{validationHdf5Filename}
\title{Create hdf5 filename from row and col for the validation}
\usage{
validationHdf5Filename(filedir, platename, row, col, configdir = "NONE")
}
\arguments{
\item{filedir}{The directory of the file}

\item{platename}{The plate that the well is on}

\item{row}{The row of the well}

\item{col}{The column of the well}

\item{configdir}{This parameter is useless but still here for backwards compatibility}
}
\value{
The HDF5 filename
}
\description{
Creates the hdf5 filename with given parameters for the hdf5 validation
}
\examples{
print(validationHdf5Filename)
}
\author{
Jan Sauer
}
