% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{watchdogStart}
\alias{watchdogStart}
\title{Start the watchdog}
\usage{
watchdogStart(configdir)
}
\arguments{
\item{configdir}{The configuration file, which defines all the directories, the number of z-stacks, channels, and fields}
}
\value{
Nothing is returned
}
\description{
Starts the watchdog. Before starting the watchdog the first time, you have to set it up with setupWatchdog. Note that you have to call watchdogStop before restarting it after a crash.
}
\examples{
print(watchdogStart)
}
\author{
Bernd Fischer
}
