% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.R
\name{watchdogStop}
\alias{watchdogStop}
\title{Stop the watchdog}
\usage{
watchdogStop(configdir)
}
\arguments{
\item{configdir}{The configuration file, which defines all the directories, the number of z-stacks, channels, and fields}
}
\value{
Nothing is returned
}
\description{
This function kills the process of the R session running the watchdog. You have to call watchdogStop before restarting it after a crash.
}
\examples{
print(watchdogStop)
}
\author{
Bernd Fischer
}
