% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{wells}
\alias{wells}
\title{Returns a data.frame with row names, col names and well names for the specified mult-well plate}
\usage{
wells(nrWells = 384)
}
\arguments{
\item{nrWells}{The design of the multi-well plate. Either 96, 384 (default), or 1536}
}
\value{
A data.frame with nrWells rows with well names, row names and col names.
}
\description{
This function returns a data.frame with well names, row names, and col names for the specified
multi-well plate. Currently supported are 96-well plates, 384 well plates, and 1536 well plates
}
\examples{
print(wells)
}
\author{
Bernd Fischer
}
