% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\docType{data}
\name{ecg}
\alias{ecg}
\title{PROMISE ECG data}
\format{a \code{data.table} with 125462 observations on the following 23 variables:
\describe{
\item{\code{MRN}}{character Medical record number}
\item{\code{ECGDate}}{POSIXct ECG exam date, YYYY-MM-DD HH:MM:SS}
\item{\code{PRInterval}}{numeric Interval between start of P wave and start QRS, ms}
\item{\code{QRSDur}}{numeric Duration of the QRS-complex, ms}
\item{\code{QTCorrected}}{numeric Corrected QT-interval according to Bazzet's formula, ms}
\item{\code{AXIS}}{numeric R wave axis, degree}
\item{\code{original_ecg_statement}}{character Original diagnosis statement in raw ECG XML}
\item{\code{filepath}}{character Name of the raw ECG XML file}
\item{\code{Normal_conduction}}{logical}
\item{\code{Sinus_rhytm}}{numeric}
\item{\code{SVT}}{numeric Supraventricular tachycardia (see details)}
\item{\code{Junctional_rhythm}}{numeric Rythm originating from AV node or His Bundle}
\item{\code{LBBB}}{numeric Left Bundle Branch Block, conventional criteria}
\item{\code{RBBB}}{numeric Right Bundle Branch Block}
\item{\code{RBBB_LAFB}}{logical Bifascicular Block, RBBB + LAFB}
\item{\code{RBBB_LPFB}}{logical Bifascicular Block, RBBB + LPFB}
\item{\code{IVCD}}{numeric Non-specific Intraventricular Conduction Delay}
\item{\code{LAFB}}{numeric Left Anterior Fascicular Block}
\item{\code{LPFB}}{numeric Left Posterior Fascicular Block}
\item{\code{AFIB}}{numeric Atrial Fibrillation}
\item{\code{PACE}}{numeric Paced rhytm, eg. pacemaker}
\item{\code{WPW}}{numeric Wolf-Parkinson White, prexcitation}
\item{\code{Exclude}}{numeric Diagnosis statement indicative for exclusion, (see details)}
}}
\source{
The data comes from Duke University Hospital, Durham, NC, USA
}
\usage{
ecg
}
\description{
PROMISE ECG data
}
\details{
Date of data extraction 2017-03-31

\describe{
\item{Normal_conduction}{See end \code{vignette("promise", package = "promise")} for logic}
\item{SVT}{= 1 if ecg diagnosis statement during parsing contained:
\itemize{
\item{Atrial reentry tachycardia}
\item{SVT}
\item{Flutter}
\item{Supra ventricular tachycardia}
\item{Atrial tachycardia}
}}
\item{Exclude}{= 1 if ecg diagnosis statement during parsing contained:
\itemize{
\item{strip}
\item{lvad}
\item{idioventricular rhythm}
\item{ventricular tach}
\item{isorhyth}
\item{ventricular escape}
\item{asys}
\item{ventricular rhyt}
\item{atrial lead}
\item{stemi}
}
}
}
}
\seealso{
\link{promise}
\link{promiseData}
}
\author{
Daniel Loewenstein, \email{loewenstein.daniel@gmail.com}
}
\keyword{datasets}
