% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_api.R
\name{loadFeatures}
\alias{loadFeatures}
\title{Get feature vector by compound}
\usage{
loadFeatures(platename, configdir)
}
\arguments{
\item{platename}{The plate ID}

\item{configdir}{The directory of the configuration files}
}
\value{
A data.frame of features. Each row represents one well's field, 
        each column represents a feature.
}
\description{
Retrieve all feature for a plate
}
\examples{
print(loadFeatures)
}
\author{
Jan Sauer
}
