% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\docType{package}
\name{promise}
\alias{promise}
\alias{promise-package}
\title{Package Title}
\description{
Datapackage for demographics and ecg data derived from Duke University Hospital 2002-2017.
}
\details{
promise
A data package for the PROMISE database.

Use \code{data(package='promise')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{promiseData}
\link{ecg}
}
\author{
Daniel Loewenstein, \email{loewenstein.daniel@gmail.com}, Karin Johansson, \email{karin.g.johansson@gmail.com}
}
