% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_path.R
\name{set_data_path}
\alias{set_data_path}
\title{Set promise data path}
\usage{
set_data_path(path, .update = FALSE)
}
\arguments{
\item{path}{A \code{character} path to and including name and extension of the .rda file.}

\item{.update}{A \code{logical}, for whether or not to update symlink.}
}
\description{
Setting up symlink for promise data
}
\details{
Since promise as of date contains PHI data processed data can't be bundled with
the package. If you mount the box.com folder as a local drive, one can symlink
the data to the promise data folder.
}
